refactor: updated schema import paths for plugins
