+++ imports/plugins/included/product-variant/containers/variantFormContainer.js
@@ -9,7 +9,7 @@ import { ReactionProduct } from "/lib/api";
 import { Packages } from "/lib/collections";
 import { Reaction, i18next } from "/client/api";
 import { TaxCodes } from "/imports/plugins/core/taxes/lib/collections";
-import { ProductVariant } from "/lib/collections/schemas/products";
+import { ProductVariant } from "/lib/collections/schemas";
 import VariantForm from "../components/variantForm";
 
 const wrapComponent = (Comp) => (
